+++ noodles-cram/src/async/reader/record.rs
@@ -631,7 +631,6 @@ where
             &mut self.external_data_readers,
         )
         .await
-        .map(char::from)
         .and_then(|id| {
             feature::Code::try_from(id).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
         })