+++ noodles-cram/src/record/feature/code.rs
@@ -31,12 +31,49 @@ pub enum Code {
     HardClip,
 }
 
+/// An error returned when a byte fails to convert to a feature kind.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct TryFromByteError(u8);
+
+impl error::Error for TryFromByteError {}
+
+impl fmt::Display for TryFromByteError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "invalid code: {:#x}", self.0)
+    }
+}
+
 /// An error returned when a character fails to convert to a feature kind.
+#[deprecated(since = "0.13.0", note = "Use `TryFromByteError` instead.")]
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct TryFromCharError(char);
 
+impl TryFrom<u8> for Code {
+    type Error = TryFromByteError;
+
+    fn try_from(n: u8) -> Result<Self, Self::Error> {
+        match n {
+            b'b' => Ok(Self::Bases),
+            b'q' => Ok(Self::Scores),
+            b'B' => Ok(Self::ReadBase),
+            b'X' => Ok(Self::Substitution),
+            b'I' => Ok(Self::Insertion),
+            b'D' => Ok(Self::Deletion),
+            b'i' => Ok(Self::InsertBase),
+            b'Q' => Ok(Self::QualityScore),
+            b'N' => Ok(Self::ReferenceSkip),
+            b'S' => Ok(Self::SoftClip),
+            b'P' => Ok(Self::Padding),
+            b'H' => Ok(Self::HardClip),
+            _ => Err(TryFromByteError(n)),
+        }
+    }
+}
+
+#[allow(deprecated)]
 impl error::Error for TryFromCharError {}
 
+#[allow(deprecated)]
 impl fmt::Display for TryFromCharError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
@@ -47,9 +84,12 @@ impl fmt::Display for TryFromCharError {
     }
 }
 
+#[allow(deprecated)]
 impl TryFrom<char> for Code {
     type Error = TryFromCharError;
 
+    #[allow(useless_deprecated)]
+    #[deprecated(since = "0.13.0", note = "Use `TryFrom<u8>` instead.")]
     fn try_from(c: char) -> Result<Self, Self::Error> {
         match c {
             'b' => Ok(Self::Bases),
@@ -92,6 +132,24 @@ impl From<Code> for char {
 mod tests {
     use super::*;
 
+    #[test]
+    fn test_try_from_u8_for_code() {
+        assert_eq!(Code::try_from(b'b'), Ok(Code::Bases));
+        assert_eq!(Code::try_from(b'q'), Ok(Code::Scores));
+        assert_eq!(Code::try_from(b'B'), Ok(Code::ReadBase));
+        assert_eq!(Code::try_from(b'X'), Ok(Code::Substitution));
+        assert_eq!(Code::try_from(b'I'), Ok(Code::Insertion));
+        assert_eq!(Code::try_from(b'D'), Ok(Code::Deletion));
+        assert_eq!(Code::try_from(b'i'), Ok(Code::InsertBase));
+        assert_eq!(Code::try_from(b'Q'), Ok(Code::QualityScore));
+        assert_eq!(Code::try_from(b'N'), Ok(Code::ReferenceSkip));
+        assert_eq!(Code::try_from(b'S'), Ok(Code::SoftClip));
+        assert_eq!(Code::try_from(b'P'), Ok(Code::Padding));
+        assert_eq!(Code::try_from(b'H'), Ok(Code::HardClip));
+        assert_eq!(Code::try_from(b'Z'), Err(TryFromByteError(b'Z')));
+    }
+
+    #[allow(deprecated)]
     #[test]
     fn test_try_from_char_for_code() {
         assert_eq!(Code::try_from('b'), Ok(Code::Bases));